Ajustes en la informacion del comic en pantalla completa
