+++ src/components/Buscador.js
@@ -3,7 +3,7 @@ import { getCharactersByNameAccion } from "../APIS/MarvelAPI";
 import { getComicsByNameAccion } from "../APIS/ComicsAPI";
 import { useDispatch } from "react-redux";
 import loupe from "../images/loupe.svg";
-import "./Buscador.css";
+import "../styles/Buscador.css";
 
 const Buscador = () => {
   const [value, setValue] = useState("");
@@ -18,7 +18,6 @@ const Buscador = () => {
 
   const handleSubmit = (e) => {
     e.preventDefault();
-    console.log({ value });
     dispatch(getCharactersByNameAccion(value));   //Busca por personajes
     dispatch(getComicsByNameAccion(value));       //Busca por comics
