+++ src/styles/Buscador.css
