+++ src/styles/ComicsFull.css
@@ -10,8 +10,7 @@
   box-shadow: 4px 3px 15px 0 rgb(0 0 0 / 20%), 0 15px 20px 0 rgb(0 0 0 / 19%);
   position: absolute;
   }
-  
-  .imgComiC {
+ .imgComiC {
     width: 60%;
     height: 100%;
     vertical-align: middle;
@@ -19,12 +18,12 @@
     overflow: hidden;
     display: contents;
   }
-  .imgComiC img {
+.imgComiC img {
     width: auto;
   }
   .infoComiC {
     width: 75%;
-    margin: 53px;
+    margin: 18px 53px;
     overflow: auto;
     font-size: 16px;
   }
@@ -32,7 +31,30 @@
     margin-bottom: 5px;
     font-size: 24px;
     padding-bottom: 40px;
+    padding-top: 20px;
   }
+.infoComiC a {
+    padding: 10px;
+    background-color: #f11212;
+    text-decoration: none;
+    color: white;
+    font-weight: 600;
+  }
+  
+  .infoButtons {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+  }
+  .infoButtons p {
+    margin-top: 0px;
+    margin-left: 20px;
+    
+
+  }
+ 
+  
+
   .infotext {
     margin: 10px 0;
     font-weight: 600;
@@ -44,13 +66,7 @@
     padding: 38px 0;
     line-height: 28px;
   }
-  .infoComiC a {
-    padding: 10px;
-    background-color: #f11212;
-    text-decoration: none;
-    color: white;
-    font-weight: 600;
-  }
+  
   .back {
     padding: 10px;
     width: 102px;
