+++ src/styles/Galeria.css
