+++ src/utils/Constants.js
@@ -1,9 +1,3 @@
-export const OBTENER_CHARACTERS_EXITO = "OBTENER_CHARACTERS_EXITO";
-export const SIGUIENTE_CHARACTERS_EXITO = "SIGUIENTE_CHARACTERS_EXITO";
-export const OBTENER_CHARACTER_POR_NOMBRE_EXITO = "OBTENER_CHARACTER_POR_NOMBRE_EXITO";
-export const ANTERIOR_CHARACTER_EXITO = "ANTERIOR_CHARACTER_EXITO";
-export const SET_WAIT = "SET_WAIT";
-export const SHOW_BUTTONS = "SHOW_BUTTONS";
 
 
 
@@ -12,3 +6,8 @@ export const verify = (value) => {
         return value
 
 }
+
+export const cap = (s) => {
+    if (typeof s !== 'string') return ''
+    return s.charAt(0).toUpperCase() + s.slice(1)
+  }
